+++ lib/widgets/common/app_bar/crumb_line.dart
@@ -1,5 +1,8 @@
+import 'dart:math';
+
 import 'package:aves/model/settings/settings.dart';
 import 'package:aves/theme/icons.dart';
+import 'package:aves/widgets/common/identity/aves_filter_chip.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -7,20 +10,22 @@ class CrumbLine<T> extends StatefulWidget {
   final List<String> Function(BuildContext context) split;
   final T Function(BuildContext context, int index) combine;
   final void Function(T combined) onTap;
+  final WidgetBuilder? lastCrumbBuilder;
 
-  static const EdgeInsets padding = EdgeInsets.only(top: 6, bottom: 20);
+  static const EdgeInsets padding = EdgeInsets.only(top: 4, bottom: 8);
 
   const CrumbLine({
     super.key,
     required this.split,
     required this.combine,
     required this.onTap,
+    required this.lastCrumbBuilder,
   });
 
   @override
   State<CrumbLine<T>> createState() => _CrumbLineState<T>();
 
-  static double getPreferredHeight(TextScaler textScaler) => textScaler.scale(22) + padding.vertical;
+  static double getPreferredHeight(TextScaler textScaler) => max(AvesFilterChip.minChipHeight, textScaler.scale(22)) + padding.vertical;
 }
 
 class _CrumbLineState<T> extends State<CrumbLine<T>> {
@@ -72,12 +77,14 @@ class _CrumbLineState<T> extends State<CrumbLine<T>> {
         final text = parts[index];
         if (index == parts.length - 1) {
           return Center(
-            child: DefaultTextStyle.merge(
-              style: TextStyle(
-                color: Theme.of(context).colorScheme.primary,
-              ),
-              child: _buildText(text),
-            ),
+            child:
+                widget.lastCrumbBuilder?.call(context) ??
+                DefaultTextStyle.merge(
+                  style: TextStyle(
+                    color: Theme.of(context).colorScheme.primary,
+                  ),
+                  child: _buildText(text),
+                ),
           );
         }
         return GestureDetector(